Pagination of stream complete with working images
To include an image in the stream, you have to use the streamimage
shortcode, e.g.

{{< streamimage filename="bobbins.jpg" >}}

bobbins.jpg must be saved in content/stream/img.

You can include a "size" param in the shortcode (a string as number),
which is translated as a height attribute in the img html tag.